+++ prisma/zod/oms_product.ts
@@ -5,7 +5,7 @@ export const oms_ProductModel = z.object({
   id: z.string().optional(),
   name: z.string().min(1, { message: "Name is required" }),
   description: z.string().min(1, { message: "Description is required" }),
-  price: z.number().positive({ message: "Price must be greater than 0" }),
+  price: z.number().int().positive({ message: "Price must be greater than 0" }),
   stock: z.number().int().int({ message: "Stock is required" }).positive({ message: "Stock must be greater than 0" }),
   createdAt: z.date().optional(),
   updatedAt: z.date().optional(),
